feat: implement DropdownContent component for user profile interactions
