+++ src/components/footer/FooterTypes.ts
@@ -1,0 +1,10 @@
+export type FooterLink = {
+  path: string;
+  label: string;
+};
+
+export type FooterLinksProps = {
+  title: string;
+  links: FooterLink[];
+  className?: string;
+};
