+++ src/components/header/NavLinks.tsx
@@ -6,8 +6,10 @@ import styles from "../../styles/components/common/Header.module.scss";
 
 const NavLinks = () => (
   <ul className={styles.navList}>
+    {/* Map through the nav links from the config */}
     {headerLinks.navLinks.map((item) => (
       <li key={item.path}>
+        {/* Render NavLink for each nav link */}
         <NavLink
           to={item.path}
           className={({ isActive }) => clsx(isActive && styles.active)}
